refactor: web/console 구분하여 패키징
