+++ src/main/java/chess/ConsoleChessApplication.java
@@ -1,8 +1,8 @@
 package chess;
 
-import chess.controller.ChessController;
+import chess.controller.console.ChessController;
 
-public class Application {
+public class ConsoleChessApplication {
     public static void main(String[] args) {
         ChessController chessController = new ChessController();
         chessController.run();
