+++ src/main/java/chess/controller/web/WebChessController.java
@@ -1,18 +1,18 @@
-package chess.controller;
+package chess.controller.web;
 
 import chess.domain.game.BoardFactory;
 import chess.domain.game.Command;
 import chess.domain.game.Game;
 import chess.domain.piece.Color;
-import chess.dto.BoardDTO;
-import chess.dto.ColorDTO;
-import chess.dto.ScoreDTO;
+import chess.controller.web.dto.BoardDTO;
+import chess.controller.web.dto.ColorDTO;
+import chess.controller.web.dto.ScoreDTO;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-public class ChessControllerForUI {
+public class WebChessController {
     Game game;
 
     public void init() {
